Show an estimated time-left indicator for par2-repair and unrar
